Combine DeObfuscator into AssemblyUtils
